rare-event-detection/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
diffusers==0.24.0
accelerate==0.25.0
numpy==1.24.3
opencv-python==4.8.1.78
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0